requirements.txt
openai
python-dotenv
moviepy
pydub